Lumberyard Release 1.12.0.0
        Binaries Uncompressed Size: 18.45GB (19807604685 bytes)

+++ dev/Code/Framework/AzCore/AzCore/Math/Guid.h
@@ -32,7 +32,7 @@ typedef struct _GUID {
     _GUID() = default;
 #endif // QUUID_H
 
-    unsigned long  Data1;
+    uint32_t       Data1;
     unsigned short Data2;
     unsigned short Data3;
     unsigned char  Data4[ 8 ];
